Khonshu v1/requirements.txt
pyautogui
pygame
pynput
obs-websocket-py
pydirectinput
google-generativeai
pyaudio
SpeechRecognition
vosk
comtypes
pycaw